rename internal/nodes to player; network cleanup;
