+++ internal/packages/player/interfaces/interfaces.go
